// shared-logic/model-user/model-user-urls.d.ts
export declare class ModelUserUrls {
    static DTO_CREATE_URL_MODEL_USER_ADMINS: string;
    static DTO_UPDATE_URL_MODEL_USER_ADMINS: string;
    static DTO_DELETE_URL_MODEL_USER_ADMINS: string;
    static DTO_UPDATE_URL_MODEL_USER_MEMBER: string;
    static DTO_DELETE_URL_MODEL_USER_MEMBER_IMAGES: string;
    static MAIN_URL: string;
    /** /zzzModelUsers/{id} */
    static URL_MODEL_USERS: string;
    /** /zzzModelUserAdmins/{id} */
    static URL_MODEL_USER_ADMINS: string;
    /** /zzzModelUserMembers/{id} */
    static URL_MODEL_USER_MEMBERS_V2: string;
    /** status/{uid}/model */
    static USER_PRESENCE: string;
    /** /aaaModelUserLog{id} */
    static MODEL_USER_LOG: string;
    /** /zzzUserPresenceLog/{timestamp}  */
    static USER_PRESENCE_LOG: string;
    /** /refUserPresenceLog/id/timestamp */
    static REF_USER_PRESENCE_LOG: string;
    /** /refUserActivityNotification/{idUser}/{idDto}/{timestamp: number, msg: string, longMsg: string} */
    static REF_USER_ACTIVITY_NOTIFICATION: string;
    private constructor();
}